+++ CHANGELOG.md
@@ -7,10 +7,12 @@ This version drops support for python 3.9 and below, and ONE is now in remote mo
 - supports python >= 3.10 only
 - OneAlyx uses remote mode by default, instead of auto
 - OneAlyx.search now updates the cache tables in remote mode as paginated sessions are accessed
+- Datasets table file_size column nullable by default
 
 ### Added
 
 - one.alf.cache.remove_cache_table_files and One.\_remove_cache_table_files for deleting cache table files
+- one.alf.cache.EMPTY_DATASETS_FRAME and EMPTY_SESSION_FRAME vars for table column, index, and dtype template
 
 ## [2.11.2]